Simplifies memory management with clearer move semantics
With Zig, doing something like:

    a.* = a.*.eval();

can cause a memory leak of `a`'s data, which hasn't been freed.
The way to solve this is to free the old value before reassigning it to `a`:

    const result = a.*.eval();
    a.*.free(allocator); // defer or errdefer shouldn't be used for moves
    a.* = result;

However, in the actual code, the result of `eval` can be the same
`a` pointer passed in! So the fixed example is:

    const result = a.*.eval();
    if (result != a.*) a.*.free(allocator);
    a.* = result;